+++ app/routes/defaultRoutes.js
@@ -1,0 +1,19 @@
+const router = require('express').Router();
+
+// Default routes pels errors
+
+router.get("*", (req, res) => {
+    res.status(404)
+    res.send({
+        error: 'Route not found'
+    });   
+});
+
+router.post("*", (req, res) => {
+    res.status(404)
+    res.send({
+        error: 'Route not found'
+    });   
+});
+
+module.exports = router
